+++ R/mst.R
@@ -1,7 +1,6 @@
-#' Edges of minimum spanning tree
+#' Minimum spanning tree
 #'
-#' Calculate or plot the ends of each edge of the minimum spanning tree of a
-#' distance matrix.
+#' Calculate or plot the minimum spanning tree of a distance matrix.
 #'
 #' @param distances Either a matrix that can be interpreted as a distance
 #' matrix, or an object of class `dist`.
@@ -12,11 +11,13 @@
 #' @param \dots Additional parameters to send to `[lines()]`.
 #'
 #' @return `MSTEdges()` returns a matrix in which each row corresponds to an
-#' edge of the minimum spanning tree, and each column lists the index of the
-#' entry in `distances` at which the line begins and ends.
+#' edge of the minimum spanning tree, listed in non-decreasing order of length.
+#' The two columns contain the indices of the entries in `distances` that
+#' each edge connects, with the lower value listed first.
 #'
 #' @seealso
-#' Calculate minimum spanning tree: [`ape::mst()`].
+#' Slow implementation returning the association matrix of the minimum spanning
+#' tree: [`ape::mst()`].
 #'
 #' @references
 #' \insertRef{Gower1969}{TreeTools}
@@ -27,20 +28,15 @@
 #'                    0, 2, 0, 2, 1, 1.1,
 #'                    0, 0, 0, 0, 1, -1), 6)
 #' distances <- dist(points)
-#' MSTEdges(distances)
+#' mst <- MSTEdges(distances)
+#' MSTLength(distances, mst)
 #' plot(points[, 1:2], ann = FALSE, asp = 1)
 #' MSTEdges(distances, TRUE, x = points[, 1], y = points[, 2], lwd = 2)
 #' @template MRS
-#' @importFrom ape mst
 #' @importFrom graphics lines
 #' @export
 MSTEdges <- function (distances, plot = FALSE, x = NULL, y = NULL, ...) {
-  umst <- ape::mst(distances)
-  edges <- umst == 1L
-  from <- umst
-  from[edges] <- unlist(apply(edges, 1, which))
-  to <- t(from)
-  ends <- cbind(from[lower.tri(from) & edges], to[lower.tri(to) & edges])
+  ends <- MinimumSpanningTree(distances)
   if (plot) {
     apply(ends, 1, function (edge)
       lines(x[edge], y[edge], ...))
@@ -48,4 +44,26 @@ MSTEdges <- function (distances, plot = FALSE, x = NULL, y = NULL, ...) {
   } else {
     ends
   }
+}
+
+#' @rdname MSTEdges
+#' @param mst Optional parameter specifying the minimum spanning tree in the
+#' format returned by `MSTEdges()`; if `NULL`, calculated from `distances`.
+#' @return `MSTLength()` returns the length of the minimum spanning tree.
+#' @export
+MSTLength <- function (distances, mst = NULL) {
+  distMat <- as.matrix(distances)
+  if (is.null(mst)) mst <- MSTEdges(distances)
+  sum(apply(mst, 1L, function (x) distMat[x[1], x[2]]))
+}
+
+MinimumSpanningTree <- function(distances) UseMethod("MinimumSpanningTree")
+
+MinimumSpanningTree.dist <- function (distances) {
+  minimum_spanning_tree(order(distances, decreasing = TRUE) - 1L)
+}
+
+MinimumSpanningTree.matrix <- function (distances) {
+  dists <- distances[lower.tri(distances)]
+  minimum_spanning_tree(order(dists, decreasing = TRUE) - 1L)
 }