+++ tests/testthat/test-mst.R
@@ -5,8 +5,12 @@ test_that('MST edges calculated correctly', {
                      0.1, 2,   0, 2, 1, 1.1,
                        0, 0,   0, 0, 1, -1), 6)
   distances <- dist(points)
-  expect_equal(matrix(c(5, 6, 6, 5, 5, 1, 1:4), 5),
-               MSTEdges(distances, FALSE))
+  apeMst <- matrix(c(5, 6, 6, 5, 5, 1, 1:4), 5)
+  distMat <- as.matrix(distances)
+  expect_equal(MSTLength(distances, apeMst),
+               MSTLength(distances))
+  expect_equal(MSTLength(distances),
+               MSTLength(distMat))
   MSTPlot <- function () {
     plot(points, asp = 1, ann = FALSE)
     expect_equal(MSTEdges(distances, FALSE),